cros-ide: Add product icon
The product icon will be shown on the public marketplace.
I converted svg to png using inkscape.

It seems we need to update the repository url so that the icon is
properly shown in the marketplace (according to the warning shown on
VSCode "An icon requires a repository with HTTPS protocol to be
specified in this package.json."), but it can be done in a separate
patch.

BUG=b:216583946
TEST=`./install.sh --dev` and open extension page https://screenshot.googleplex.com/BbLa5aNBDfBDStv

Change-Id: I02427022f97c391f94ae2b184a6c50f882a35e55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3666547
Reviewed-by: Eric Lok <lokeric@google.com>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>