cros-ide: remove welcome view for no boards
Since we now support host packages, we will not have a situation when
there are no boards. We are therefore removing a welcome view for this
case. The information how to set up boards will still be shown, because
we have an extra welcome view for new users, that needs to be dismissed
before boards and packages are shown.

BUG=b:235940443
TEST=npm t

Change-Id: I6a9ea2476a76419885a3731b85fa323e6134a52d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3703807
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>