cros-ide: Run tast tests from the IDE.
There is currently no way to run tast tests from the IDE. This change
adds the feature to run a test from a tast test file, using a button
from the menu bar. To manage open SSH connection, the SshSession class
was created and a map of host names to SshSession was added to the
context. The net_utl.ts file was created to export the findOpenPort()
function, which is useful for both VNC sessions and SSH sessions.

BUG=b:211819488
TEST=ran npm test and development host
TEST=https://screenshot.googleplex.com/5UNNEradoyjTiWC
TEST=https://screenshot.googleplex.com/C2nqCikodaFcpaM
TEST=https://screenshot.googleplex.com/7gFiUsxNEt5JFZs

Change-Id: Ie0d8d06e5fbef799ca88e07dbf26e0bb88e435f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3812015
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Uwem Wilson <uwemwilson@google.com>
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Tested-by: Uwem Wilson <uwemwilson@google.com>