cros-ide: Use job queue to throttle fetching of Gerrit comments
Before this change the code was assuming that fetching Gerrit comments
would always complete in 10 seconds, and otherwise the operations of
fetching which are triggered every 10 seconds overlap, possibly
producing burst of requests to Gerrit, ultimately producing a lot of
429 errors.

After this change, the requests are queued using JobManager; if three
or more requests are queued, intermediate requests are discarded, and
operations of fetching don't overlap.

BUG=b:270101085
TEST=npm run unit-test
TEST=Create a local branch with 50 commits all uploaded to Gerrit. Open
a file in the branch and observe the log. Confirm no 429 errors are
reported in the log.

Change-Id: I7ec08d274bd677b8607dfd29a48876baaed5b4fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4473747
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>