cros-ide: Don't rerun emerge when build fails
When the command to generate the compilation database fails, it is
usually meaningless to rerun the same command. We already instruct
the user to fix the problem and reload the IDE in the error message,
so there is no need to automatically rerun the command hoping the
problem may have been manually fixed in the meantime.

BUG=b:284799433
FIXED=b:284799433
TEST=npm t

Change-Id: Id97513927fe39e27ead5053247ad1b06197d2605
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4573262
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>