cros-ide: metrics: add code path to send metrics to GA4
It is still sending to UA but will be immediately ready to change to GA4
when the config boolean flag cros-ide.underDevelopment.metricsGA4 is
flipped (to true).

Ideally before that happens, all metrics.send callsites:
1. provides a GA4 event_name rule conforming Event.name.
2. gives any extra info as custom dimensions (not label) or metrics (not
   value).
But if the work is not completed then, as a temporary measure,
1. Event.name (or Event.description as fallback) will be sanitized to
   make sure it is a valid event_name.
2. pass them as label or value.

Utility function to format request body for GA4 is moved to
metrics_util.ts with unit tests.

BUG=b:281925148
TEST=manually verify that events are recorded on UA realtime report
TEST=manually verify that when flag is flipped, events are recorded on \
     GA4 realtime report
TEST=npm run test

Change-Id: I7f3f63e7e7fc40a2d8bdeb506b69afc3e50b569d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4584536
Auto-Submit: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>