cros-ide: Remove code for device management V2
We don't have immediate plan to polish and publish device management V2.
This change remove the code for it so to decrease the complexity of
refactoring around the device management feature such as for
b:260009987.

We can always revive the code by reverting this CL when we get time to
rework on it.

BUG=b:263322986
TEST=npm t

Change-Id: Icd360b5eeac63b1bc03cdf134105a0e252b86c59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4576582
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>