ide: Migrate build from webpack to esbuild
As a sequel of the change to build webview with esbuild, this change
lets esbuild build other part of the extension as well. It improves the
total build time about 20%.

Before this change, npm run build took about 8 seconds, but after this
change, it takes about 6 seconds.

Now that webpack is only used to run the plugins, but it still requires
an entry point, so we add an empty JS file as a fake entry point of
webpack.

BUG=b:275545000
TEST=npm t
TEST=confirm VNC view for a leased device works as before.
TEST=confirm log viewer for a leased device works as before.

Change-Id: Id0d4c5670cb968af1a397b32623e567e63eea6c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4705634
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>