ide: Improve speed of building tests
This change improves the speed of npm run build-tests from 10 seconds to
3 seconds. This change achieves the speed up by migrating how we build
tests from tsc to esbuild.

Simply substituting the build system to esbuild causes a lot of test
failures though. The problem is that we have been relying on tsc's
output being linient allowing us to mock a module's exported fields
directly from tests but it is no longer true for esbuild's output, as
it follows the spec more rigorously.
To workaround this, this change introduces an allow list of files
that are compiled separately into the form as linient as tsc's output.
This allows us to mock modules that are on the allow list from tests
that are also in the allow list.

We should eventually empty the allow list but it can be done separetely
from this change.

BUG=b:275545000
TEST=npm t
TEST=time npm run unit-test # 10 seconds, twice as fast as before

Change-Id: If5cc0f2b9fa2150675cfc70e889f544791bbf635
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4705637
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>