ide: boards and packages: add context menu to workon packages
This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload
TEST=right click on codelab, click the menu to workon it, right click
it again, and confirm the menu is now to stop working on it.
TEST=right click on codelab, click the menu to stop working on it,
right click it again, and confirm the menu is now to start work on it.
TEST=right click on dev-go/delve and confirm no menu related to workon
is shown.

Change-Id: I1f522f0621cc199538735c9d660514cfce437ed9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4835435
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>