ide: boards and packages: add a context menu to change default board
This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload
TEST=set up betty and amd64-generic boards, set betty as the default
board, confirm three boards including host appear on the tree view,
right-click each board item and confirm the menu to set as the default
board appears only on amd64-generic, and confirm by clicking the menu
item amd64-generic is set as the default board and the change is
reflected on the tree view UI.

Change-Id: I47237f98fb5c0c21718a16957bc18cc735577bc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4831691
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>