ide: Setup debug configuration
This change adds a debugConfiguration instead of `launch.json` file in Step 5-5 of go/debug-tast-tests.
There remain some issues. When opening some OUTPUT tab after starting debug tast tests, the error `Variable ${fileWorkspaceFolder}: can not find workspace folder of ...` occurs. When focusing on the file that is not you started the debug tast test, the current-file in debugTastTestTask does not set to the file which was active when the debug command was invoked.

BUG=b:255632861
TEST=npm run preupload
TEST=Manually run “Debug Test” in go file.

Change-Id: Ie24c1675223e06da858aba83426666687aae1fa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4825428
Tested-by: Chiaki Ushimaru <uchiaki@google.com>
Commit-Queue: Chiaki Ushimaru <uchiaki@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>