device: run image check before deploying package
If the check passes, nothing will be shown. The check takes ~4s
(measured over 5 attempts).
Otherwise, user will have the option to flash the device first or ignore
the warning. The deploy command will run directly after flashing
completes.
If the user cancels at any point of the check, the deploy package
command will also end.
If the user chooses to flash the image (after a failed check) and the
`cros flash` command fails, they can further choose to continue to
deploy or not.

Report result of check as metrics to understand usefulness of the
feature.

BUG=b:308059094
TEST=manually deploy a package with wrong version; `cros deploy` \
     started immediately without prompt after flashing complete
TEST=failed check reported as quickpick:
     https://screenshot.googleplex.com/7uwDbQYALgwJTCN
TEST=npm t

Change-Id: Ib0872736a1189d1a1efde5ee2efa581e08c01a3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5300479
Commit-Queue: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>