Add minimatch dependency for matching glob patterns and run npm install
For cros format feature to determine if the target file should be
formatted or not (if it is in the relevant .presubmitignore).

The version is chosen  manually to make sure glob and minimatch are
compatible.  If installed with `npm install minimatch --save-prod`, the
types used by glob will be different from those exported from that
minimatch.

BUG=b:329601336
TEST=npm t

Change-Id: I1bbce49a234ab0d6a43dd465c928b86e11471956
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5439515
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>