Forbid string literals containing 'ChromiumIDE'
They should use the extensionName() function instead to make sure it
shows as ChromeOS on cider.

Do not block ChromeOS since it is commonly used as a reference to the
actual ChromeOS product, project, team, etc. It should be safe since
all features are first implemented on the vscode platform and migrated
to cider later, so developers would not refer to the extension as
'ChromeOS' when the code is added to shared/.

When the non-user-facing usages (e.g. config, command) are also migrated
the lint rule will not be case sensitive.

BUG=b:337139768
TEST=https://screenshot.googleplex.com/5jdZPepVhRcTwiC

Change-Id: I75ebc63f995bf00c7087c67d27d8682e70086928
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5529196
Tested-by: Grace Cham <hscham@chromium.org>
Auto-Submit: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>