cros lint: disable libchrome checks on cider
It is not for free since we need to get security approval for running
the script. As only basic linting and formatting is included for M1,
disable it for now but we could add it in the future.

BUG=b:336837488
TEST=None

Change-Id: I2dbdeb74eefbbf84c4df89eaafa5326c7fde5af8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5537229
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Auto-Submit: Grace Cham <hscham@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>