gcert: Add a command to run gcert
Enable running gcert from IDE, so that later features can automatically
call the command in case commands fail due to missing certificate.

In case VSCode server is running as a deamon, the terminal on VSCode
will not be associated with any SSH session. If that's the case, we
piggy-back on existing SSH session and set the auth socket for that
session to SSH_AUTH_SOCK on running gcert. If there are multiple
candidates we explicitly ask the user which one to use.

BUG=b:265155874
TEST=npm t
TEST=Run gcertdestroy, run the "Run gcert" command from IDE, enter
passsword and touch the key, and confirm the command succeeds.
TEST=Follow the workflow https://docs.google.com/document/d/1S72AIWofwvYVC3PLbfHZdpLdjquH1ktbZOF-4oN-CjQ/edit?resourcekey=0-kd8ku9CVHk-KY8DznTKbwA&tab=t.0#bookmark=id.8002ma9qv8ku
to set up background vscode serve-web service, create two SSH sessions
to the workstation, run the "Run gcert" command from IDE, confirm it
prompts two cands for SSH_AUTH_SOCK, select one of them, enter password
and touch the key, and confirm the command succeeds.

Change-Id: I196b47cbabbb7ad102be9e979f00239e9f71d8d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5553185
Reviewed-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>