Fix potentially flaky test for BoardsAndPackages
The fake TreeView class we provide in injected_modules has the same APIs
as VSCode's TreeView and mimics its behavior. Especially it listens to
the `onDidChangeTreeData` event `vscode.TreeDataProvider` provides, but
there's no API to wait until its handling finishes.

Before this change, the test
'Boards and packages sorts packages by favorite and workon status'
has been potentially flaky, because the following events can happen in
order.

1. In the test suite set up, it sets up fakeExec, that handles `sudo`
   commands that `getChildren` in our implementation of
   `vscode.TreeDataProvider` invokes.
2. In the test spec, it updates the status of the model backing the
   tree data provider via commands such as
   'chromiumide.boardsAndPackages.favoriteAdd', and the tree data
   provider fires events from `onDidChangeTreeData`, but as
   mentioned there's no way to tell when the handling of the events
   by TreeView finishes.
3. The test finishes and the fakeExec set up in #1 is disposed of.
4. The handler in TreeView is still running and calls `getChildren`,
   and in turn it calls `sudo` commands in real `exec`.

The potential flakiness hasn't been manifested as a real failure until
https://crrev.com/c/5568924 has been submitted, which changes the
environment variables passed to sudo in `sudo.ts`, but the fundamental
problem is that the test gets real sudo commands run.

This change fixes the issue by introducing an event in our tree data
provider impl that is fired when the class is disposed of and the number
of in-flight async operations is zero, and ensuring the test to finish
after receiving the event. This way, we can ensure all the sudo commands
are run before the test suite exits and handled by the fakeExec it
installed.

BUG=b:343350889
FIXED=b:343350889
TEST=cherry-pick https://crrev.com/c/5568924 and run `npm run unit-test`
repeatedly.

Change-Id: I76d1daf34c5724e5e97714921c9694fbd650e0e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5578980
Reviewed-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>