driver/cros: takes input path as optional parameter
The path is not used on cider since it only needs to know the cog
workspace path. This allows activating features without a specific path
(e.g. file opened or workspace folder).

This is no-op for the vscode extension. Path is always given currently,
and in new features to be added for cider where path is undefined they
will return undefined (no chroot/ not in CrOS repo).

This is for finding depot_tools path (for running `cros format`).

BUG=None
TEST=None

Change-Id: If566c85728bb7ab62c5ee31d9ad36823bcca6f58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5620276
Auto-Submit: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>