lint: Honor PRESUBMIT.cfg on running cros lint
The path to the target file must be specified with the relative path from
the repository root so that --exclude and --include paths are honored.
For constructing the relative path, add driver.path.relative and use it.
The cwd is modified accordingly.

BUG=b:337138396
FIXED=b:337138396
TEST=Open infra/recipes as a workspace, open test_plan_filtering.py,
edit a `"""` to `"`, and confirm cros lint warning is shown. Then open
recipes.py, and confirm warnings from cros lint are not shown.
TEST=npm t

Change-Id: Ia84e40d7231800e0f731100487da83a9dc0999a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5623874
Reviewed-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>