Run unit tests in parallel on npm t and CQ
Use unit-tests-parallel instead of unit-tests for npm test command.
This speeds up `npm t`. To maximize the benefit of parallelization
wireit makes, update --process=auto to --process=32; otherwise
--process=auto may use as much core as possible and blocks other npm
commands otherwise could run in parallel.

cq-test is updated to *additionally* run unit-test-parallel. Because
unit-test command is still useful for using `fdescribe` or `fit` that
are unsupported in parallel mode for local development, and we should
ensure the command works as well.

Incidentally, our current bottleneck is the integration-test command,
which on its own takes 23.2s. There seems no low hanging fruits to get
further speed up for the command, because what's slow is not the test
suites but the set up and the tear down of real VSCode the integration
tests inherently require.

BUG=b:346849057
FIXED=b:346849057
TEST=time npm t twice # The second run finishes in 24.1s. Previously it
was 42s, unit-test being the bottleneck.

Change-Id: I6816937f8d52c6b7b90f6f84ce4ffe1e84ceb9ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5632517
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>