ide: java/ls: Import java-language-server
This patch imports the latest snapshot of java-language-server source
code. It is built in parallel to the extension and included in the VSIX.

We now need sufficiently recent OpenJDK and Maven to build the
extension. We have to update the automated prerelease bot to prepare
them.

BUG=b:383450257
TEST=npm run build
TEST=npx vsce ls
TEST=./run_presubmit.sh --docker

Change-Id: I91b3ba704859db868d4218716bea8acc1183d79e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/6090778
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>